Add files via upload
You're working for an online marketplace. 
Users often search for products using keywords. Given 
a list of product names and their corresponding IDs, 
design a function that allows users to input a keyword 
and quickly find the product they are looking for. 
Implement a linear search algorithm to search for the 
keyword in the list of product names and return the 
corresponding product ID